refactor(registry): split resolver facade
